+++ core/web/presenters/job_test.go
@@ -89,6 +89,7 @@ func TestJob(t *testing.T) {
 						"offChainReportingOracleSpec": null,
 						"fluxMonitorSpec": null,
 						"keeperSpec": null,
+						"webSpec": null,
 						"errors": []
 					}
 				}
@@ -149,6 +150,7 @@ func TestJob(t *testing.T) {
 						"offChainReportingOracleSpec": null,
 						"directRequestSpec": null,
 						"keeperSpec": null,
+						"webSpec": null,
 						"errors": []
 					}
 				}
@@ -214,6 +216,7 @@ func TestJob(t *testing.T) {
 						"fluxMonitorSpec": null,
 						"directRequestSpec": null,
 						"keeperSpec": null,
+						"webSpec": null,
 						"errors": []
 					}
 				}
@@ -260,11 +263,56 @@ func TestJob(t *testing.T) {
 						},
 						"fluxMonitorSpec": null,
 						"directRequestSpec": null,
+						"webSpec": null,
 						"offChainReportingOracleSpec": null,
 						"errors": []
 					}
 				}
 			}`, contractAddress, fromAddress),
+		},
+		{
+			name: "web spec",
+			job: job.Job{
+				ID: 1,
+				WebSpec: &job.WebSpec{
+					CreatedAt: timestamp,
+					UpdatedAt: timestamp,
+				},
+				PipelineSpec: &pipeline.Spec{
+					ID:           1,
+					DotDagSource: "",
+				},
+				Type:            job.Type("web"),
+				SchemaVersion:   1,
+				Name:            null.StringFrom("test"),
+				MaxTaskDuration: models.Interval(1 * time.Minute),
+			},
+			want: `
+			{
+				"data":{
+					"type":"jobs",
+					"id":"1",
+					"attributes":{
+						"name": "test",
+						"schemaVersion": 1,
+						"type": "web",
+						"maxTaskDuration": "1m0s",
+						"pipelineSpec": {
+							"id": 1,
+							"dotDagSource": ""
+						},
+						"webSpec": {
+							"createdAt":"2000-01-01T00:00:00Z",
+							"updatedAt":"2000-01-01T00:00:00Z"
+						},
+						"fluxMonitorSpec": null,
+						"directRequestSpec": null,
+						"keeperSpec": null,
+						"offChainReportingOracleSpec": null,
+						"errors": []
+					}
+				}
+			}`,
 		},
 		{
 			name: "with errors",
@@ -317,6 +365,7 @@ func TestJob(t *testing.T) {
 						},
 						"fluxMonitorSpec": null,
 						"directRequestSpec": null,
+						"webSpec": null,
 						"offChainReportingOracleSpec": null,
 						"errors": [{
 							"id": 200,