+++ docs/CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Web job types can now be created from the operator UI as a new job. 
+
+- See example web job spec below: 
+
+```
+type            = "web"
+schemaVersion   = 1
+observationSource   = """
+ds          [type=http method=GET url="http://example.com"];
+ds_parse    [type=jsonparse path="data"];
+ds -> ds_parse;
+"""
+```
+
 - Add `MockOracle.sol` for testing contracts
 
 ### Changed